MarginCall_AzureOpenAI/requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-core
faiss-cpu
pandas
numpy
torch
scikit-learn
lightgbm
fastapi
uvicorn
python-dotenv
azure-search-documents
azure-identity
streamlit
requests
email-validator
plotly